// src/types/features/import/ImportType.ts
export interface CSVRowProps {
    SousCategory: any;
    ProductName: string;
    VariantName:string;
    Category: string;
    SubCategory?: string;
    PriceHt: string;
    Tva:string;
    PriceTtc: string;
    Stock:string;
    Brand: string;
    Manufacturer: string;
    Size?: string;
    Color?: string;
    Weight?: string;
    Laterality?: string;
    Material?: string;
    Height?: string;
    Width?: string;
    Image?: string;
}